Prepare for using standard py27 test
Instead of setting up spark in the CI job, move the set up here so that
it can be reused by other jobs as well.

Create a test-setup.sh script that developer can also use to setup their
own environment.

This makes it easier for monasca-analytics to change the setup,
documents the setup, and enables developer to reuse it.

For details see:
http://lists.openstack.org/pipermail/openstack-dev/2016-November/107784.html

Change-Id: Id1a0b8d44d5a258746dd1f443610293163e64096
Needed-By: I0fdce1a977f59cdcb8c63c2b215316cbfcf4fef4